+++ src/react.ts
@@ -4,3 +4,5 @@ export { VideoError } from "./player/video-error.js";
 export type { VideoErrorOptions } from "./player/video-error.js";
 export { useVideo } from "./player/use-video.js";
 export type { UseVideoOptions, UseVideoResult } from "./player/use-video.js";
+export { useNarration } from "./player/use-narration.js";
+export type { Narration, NarrationOptions, NarrationVoice } from "./player/use-narration.js";